todo-api/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
pymongo==4.3.3
flask-cors==3.0.10
PyJWT==2.6.0